Added chapter 2 as well as fixing some issues in chapter 1
